+++ packages/riverpod_stateful_service_generator/test/resources/success_with_positioned_parameters.dart
@@ -2,9 +2,9 @@ import 'package:riverpod/riverpod.dart';
 import 'package:riverpod_stateful_service_annotation/riverpod_stateful_service_annotation.dart';
 import 'package:stateful_service/stateful_service.dart';
 
-part 'success_with_positioned_parameter.stateful_service.dart';
+part 'success_with_positioned_parameters.stateful_service.dart';
 
 @riverpodService
-class A extends StatefulService<int> {
-  A(Ref ref, int a) : super(initialState: a);
+class A extends StatefulService<(int, int?, int)> {
+  A(Ref ref, int a, [int? b, int c = 0]) : super(initialState: (a, b, c));
 }